requirements.txt
flask
flask-cors
python-dateutil